Greatly enhanced runtime configuration change reactivity.
Before, when a setting changed, I simply recreated everything. That made GS come
to a complete halt when using the spin buttons. Now we listen to which keys have
changed and only do the changes in the UI that are necessary, which makes
everything a lot snappier. There's still problems with lag when enabling/disabling
edges frequently (maybe the better solution is to create the reactive areas and
just connect/disconnect the signals. Or toggle reactivity. I don't know enough
about clutter/gnome-shell to judge this off-hand, I have to test it. But not
today ...)